requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
typing-extensions
supabase
httpx
PyMuPDF
Pillow
torch
weasyprint
transformers
ultralytics
jinja2
langchain-core
langgraph
langchain-openai
langchain-groq